% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aem.R, R/plot.R
\name{aem}
\alias{aem}
\alias{contour.aem}
\alias{plot.element}
\alias{plot.aem}
\title{Create a analytic element model}
\usage{
aem(k, top, base, n, ...)

\method{contour}{aem}(aem, x, y, z = c("heads", "streamfunction", "potential"), asp = 1, ...)

\method{plot}{element}(element, add = FALSE, pch = 16, cex = 0.75, ...)

\method{plot}{aem}(aem, add = FALSE, xlim, ylim, frame.plot = TRUE, ...)
}
\arguments{
\item{k}{numeric, hydraulic conductivity of aquifer}

\item{top}{numeric, top elevation of aquifer}

\item{base}{numeric, base elevation of aquifer}

\item{n}{numeric, effective porosity of aquifer as a fraction of total unit volume. Used for determining flow velocities with \code{\link[=velocity]{velocity()}}.}

\item{...}{additional arguments passed to \code{\link[=plot]{plot()}}.}

\item{aem}{\code{aem} object}

\item{x}{numeric x coordinates at which the values in \code{z} are computed These must be in ascending order.}

\item{y}{numeric x coordinates at which the values in \code{z} are computed These must be in ascending order.}

\item{z}{character indicating which variable to plot. Possible values are \code{heads} (default),
\code{streamfunction}, \code{potential}, \code{Q} (magnitude of discharge vector), \code{Qx}, \code{Qy}, \code{Qz},
(x, y and z components of the discharge vector) and \code{velocity}.}

\item{asp}{the \code{y/x} aspect ratio, see \code{\link[=plot.window]{plot.window()}}. Defaults to 1 (equal unit lengths).}

\item{element}{analytic element of class \code{element} to plot. If not a point or line geometry, nothing is plotted.}

\item{add}{logical, should the plot be added to the existing plot? Defaults to \code{FALSE}.}

\item{pch}{numeric point symbol value, defaults to \code{16}.}

\item{cex}{numeric symbol size value, defaults to \code{0.75}.}

\item{xlim}{numeric, plot limits along the x-axis. Required if \code{add = FALSE}.}

\item{ylim}{numeric, plot limits along the y-axis. Required if \code{add = FALSE}.}

\item{frame.plot}{logical, should a border be drawn around the plot. Defaults to \code{TRUE}.}
}
\value{
\code{\link[=aem]{aem()}} returns an object of class \code{aem} which is a list consisting of \code{k}, \code{top}, \code{base}, \code{n},
a list containing all elements, and a logical \code{solved} indicating if the model is solved.
}
\description{
\code{\link[=aem]{aem()}} creates an analytic element model to which elements can be added

\code{\link[=contour.aem]{contour.aem()}} creates a contour plot of a computed variable, or adds the contour lines
to and existing plot.
}
\details{
When calling \code{\link[=aem]{aem()}}, if an element of class \code{headequation} is supplied, \code{\link[=solve.aem]{solve.aem()}} is called on the \code{aem}
object before it is returned.
}
\examples{
k <- 10
top <- 10
base <- 0
n <- 0.2
TR <- k * (top - base)

w <- well(xw = 50, yw = 0, Q = 200)
rf <- constant(xc = -500, yc = 0, h = 20)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = TR)

aem(k, top, base, n, w, rf, uf)
aem(k, top, base, n, list(w, rf, uf))


w <- well(xw = 50, yw = 0, Q = 200)
wi <- well(xw = 0, yw = 0, Q = -100)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = 100)
ml <- aem(k = 10, top = 10, base = 0, n = 0.2, w, wi, uf)

xg <- seq(-100, 100, length = 500)
yg <- seq(-75, 75, length = 100)

contour(ml, xg, yg, nlevels = 20, col = 'dodgerblue3', labcex = 1)
contour(ml, xg, yg, 'streamfunction', nlevels = 20, col = 'orange3', drawlabels = FALSE, add = TRUE)

ls <- linesink(x0 = -200, y0 = -150, x1 = 200, y1 = 150, sigma = 0.1)
w <- well(xw = 50, yw = 0, Q = 200)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = 100)

plot(ls)
plot(w, add = TRUE)
plot(uf) # empty

w <- well(xw = 50, yw = 0, Q = 200)
wi <- well(xw = 0, yw = 0, Q = -100)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = 100)
ls <- linesink(-75, 50, 100, 50, sigma = 1)
ml <- aem(k = 10, top = 10, base = 0, n = 0.2, w, wi, uf, ls)

plot(ml, xlim = c(-100, 100), ylim = c(-75, 75))

xg <- seq(-100, 100, length = 500)
yg <- seq(-75, 75, length = 100)

contour(ml, xg, yg, nlevels = 20, col = 'dodgerblue3')
plot(ml, add = TRUE)

}
