% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areasink.R
\name{areasink}
\alias{areasink}
\title{Create a circular area-sink analytic element with specified recharge}
\usage{
areasink(xc, yc, N, R, ...)
}
\arguments{
\item{xc}{numeric, x location of the center of the area-sink}

\item{yc}{numeric, y location of the center of the area-sink}

\item{N}{numeric, uniform constant recharge value (positive is into aquifer) in length per time.}

\item{R}{numeric, radius of the circular area-sink}

\item{...}{ignored}
}
\value{
Circular area-sink analytic element which is an object of class \code{areasink} and inherits from \code{element}.
}
\description{
\code{\link[=areasink]{areasink()}} creates a circular area-sink analytic element with constant, uniform specified recharge.
}
\examples{
areasink(xc = -500, yc = 0, N = 0.001, R = 500)

}
