% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracelines.R
\name{capzone}
\alias{capzone}
\title{Title}
\usage{
capzone(
  aem,
  well,
  time,
  npar = 30,
  dt = time/100,
  as.poly = TRUE,
  hull = FALSE,
  R = 1,
  ...
)
}
\arguments{
\item{R}{}
}
\description{
Title
}
