% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R, R/aem.R, R/areasink.R, R/element.R
\name{disc}
\alias{disc}
\alias{disc.aem}
\alias{disc.areasink}
\alias{disc.element}
\title{Calculate the complex discharge}
\usage{
disc(...)

\method{disc}{aem}(aem, x, y, as.grid = FALSE, ...)

\method{disc}{areasink}(areasink, x, y, ...)

\method{disc}{element}(element, x, y, ...)
}
\arguments{
\item{...}{ignored}

\item{aem}{\code{aem} object}

\item{x}{numeric x coordinates to evaluate \code{disc} at}

\item{y}{numeric y coordinates to evaluate \code{disc} at}

\item{as.grid}{logical, should a matrix of dimensions c(\verb{length(y), length(x)}) be returned? Defaults to \code{FALSE}.}

\item{areasink}{}

\item{element}{analytic element of class \code{element}}
}
\value{
A vector of \code{length(x)} (equal to \code{length(y)}) with the complex discharge values at \code{x} and \code{y}.
If \code{as.grid = TRUE}, a matrix of dimensions \code{c(length(y), length(x))} described by
marginal vectors \code{x} and \code{y} containing the complex discharge values at the grid points.

complex discharge influence of \code{areasink} evaluated at points \verb{x y}.
}
\description{
\code{\link[=disc]{disc()}} computes the complex discharge for an \code{aem} or \code{element} object
at the given x and y coordinates.
}
\examples{
TR <- 100
w <- well(xw = 50, yw = 0, Q = 200)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = TR)
ml <- aem(TR = 100, w, uf)

disc(ml, c(50, 0), c(25, -25))

xg <- seq(-100, 100, length = 500)
yg <- seq(-75, 75, length = 100)
disc(ml, xg, yg, as.grid = TRUE)

w <- well(xw = 50, yw = 0, Q = 200)
disc(w, c(50, 0), c(-25, 25))

}
\seealso{
\code{\link[=head.aem]{head.aem()}}, \code{\link[=omega]{omega()}}, \code{\link[=potential]{potential()}}, \code{\link[=streamfunction]{streamfunction()}}, \code{\link[=disvec]{disvec()}}
}
