% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R, R/aem.R, R/element.R
\name{discharge}
\alias{discharge}
\alias{discharge.aem}
\alias{discharge.element}
\title{Calculate the discharge vector}
\usage{
discharge(...)

\method{discharge}{aem}(aem, x, y, as.grid = FALSE, magnitude = FALSE, ...)

\method{discharge}{element}(element, x, y, as.grid = FALSE, magnitude = FALSE, ...)
}
\arguments{
\item{...}{ignored}

\item{aem}{\code{aem} object}

\item{x}{numeric x coordinates to evaluate \code{discharge} at}

\item{y}{numeric y coordinates to evaluate \code{discharge} at}

\item{as.grid}{logical, should a matrix of dimensions c(\verb{length(y), length(x)}) be returned? Defaults to \code{FALSE}.}

\item{magnitude}{logical, should the magnitude of the discharge vector be returned as well? Default to \code{FALSE}. See details.}

\item{element}{analytic element of class \code{element}}
}
\value{
A matrix of dimensions \code{c(length(x), 3)} with the \verb{Qx, Qy} and \code{Qz} components of
the discharge vector at \code{x} and \code{y}. If \code{as.grid = TRUE}, an array of dimensions
\code{c(length(y), length(x), 3)} described by marginal vectors \code{x} and \code{y} (columns and rows)
containing the \verb{Qx, Qy} and \code{Qz} components of the discharge vector.

If \code{magnitude = TRUE}, the last dimension of the returned matrix is expanded to include
the magnitude of the discharge vector, calculated as \code{sqrt(Qx^2 + Qy^2 + Qz^2)}.
}
\description{
\code{\link[=discharge]{discharge()}} computes the \verb{Qx, Qy} and \code{Qz} components of the discharge vector for an \code{aem} or \code{element} object
at the given x and y coordinates.
}
\examples{
TR <- 100
w <- well(xw = 50, yw = 0, Q = 200)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = TR)
ml <- aem(TR = 100, w, uf)

discharge(ml, c(50, 0), c(25, -25), magnitude = TRUE)

xg <- seq(-100, 100, length = 500)
yg <- seq(-75, 75, length = 100)
discharge(ml, xg, yg, as.grid = TRUE)

w <- well(xw = 50, yw = 0, Q = 200)
discharge(w, c(50, 0), c(-25, 25))
discharge(w, c(50, 0), c(-25, 25), as.grid = TRUE, magnitude = TRUE)

}
\seealso{
\code{\link[=head.aem]{head.aem()}}, \code{\link[=omega]{omega()}}, \code{\link[=potential]{potential()}}, \code{\link[=streamfunction]{streamfunction()}}, \code{\link[=discharge]{discharge()}}
}
