% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R, R/aem.R, R/element.R
\name{flow}
\alias{flow}
\alias{domega}
\alias{discharge}
\alias{darcy}
\alias{velocity}
\alias{domega.aem}
\alias{discharge.aem}
\alias{darcy.aem}
\alias{velocity.aem}
\alias{domega.element}
\alias{discharge.element}
\title{Calculate flow variables}
\usage{
domega(...)

discharge(...)

darcy(...)

velocity(...)

\method{domega}{aem}(aem, x, y, as.grid = FALSE, ...)

\method{discharge}{aem}(aem, x, y, as.grid = FALSE, magnitude = FALSE, ...)

\method{darcy}{aem}(aem, x, y, as.grid = FALSE, magnitude = FALSE, ...)

\method{velocity}{aem}(aem, x, y, as.grid = FALSE, magnitude = FALSE, R = 1, ...)

\method{domega}{element}(element, x, y, ...)

\method{discharge}{element}(element, x, y, as.grid = FALSE, magnitude = FALSE, ...)
}
\arguments{
\item{...}{ignored}

\item{aem}{\code{aem} object}

\item{x}{numeric x coordinates to evaluate at}

\item{y}{numeric y coordinates to evaluate at}

\item{as.grid}{logical, should a matrix of dimensions c(\verb{length(y), length(x)}) be returned? Defaults to \code{FALSE}.}

\item{magnitude}{logical, should the magnitude of the flow vector be returned as well? Default to \code{FALSE}. See details.}

\item{R}{numeric, retardation coefficient. Defaults to 1 (no retardation).}

\item{element}{analytic element of class \code{element}}
}
\value{
For \code{\link[=domega]{domega()}},  vector of \code{length(x)} (equal to \code{length(y)}) with the complex discharge values at \code{x} and \code{y},
If \code{as.grid = TRUE}, a matrix of dimensions \code{c(length(y), length(x))} described by
marginal vectors \code{x} and \code{y} containing the complex discharge values at the grid points.
\code{\link[=domega]{domega()}} is the derivate of \code{\link[=omega]{omega()}} in the x and y directions.

For \code{\link[=discharge]{discharge()}}, a matrix of dimensions \code{c(length(x), 3)} with the \verb{x, y} and \code{z} components (\code{Qx}, \code{Qy} and \code{Qz})
of the discharge vector at coordinates \code{x}, \code{y} and \code{z}. If \code{as.grid = TRUE}, an array of dimensions
\code{c(length(y), length(x), 3)} described by marginal vectors \code{x} and \code{y} (columns and rows)
containing the \verb{x, y} and \code{z} components of the discharge vector (\code{Qx}, \code{Qy} and \code{Qz}).

The \code{x} component of \code{\link[=discharge]{discharge()}} is the real value of \code{\link[=domega]{domega()}}, the \code{y} component
the imaginary component and \code{z} is calculated based on area-sink strengths.

If \code{magnitude = TRUE}, the last dimension of the returned matrix is expanded to include
the magnitude of the discharge/darcy/velocity vector, calculated as \code{sqrt(Qx^2 + Qy^2 + Qz^2)}
(or \code{sqrt(qx^2 + qy^2 + qz^2)} or \code{sqrt(vx^2 + vy^2 + vz^2)}, respectively).

For \code{\link[=darcy]{darcy()}}, the same as for \code{\link[=discharge]{discharge()}} but with the \code{x}, \code{y} and \code{z} components of the
Darcy flux vector (\code{qx}, \code{qy} and \code{qz}). The value are computed by dividing the values of \code{\link[=discharge]{discharge()}} by the saturated thickness.

For \code{\link[=velocity]{velocity()}}, the same as for \code{\link[=discharge]{discharge()}} but with the \code{x}, \code{y} and \code{z} components of the
average linear groundwater flow velocity vector (\code{vx}, \code{vy} and \code{vz}). The values are computed by dividing
the \code{\link[=darcy]{darcy()}} values by the effective porosity and the retardation coefficient.
}
\description{
\code{\link[=domega]{domega()}} computes the complex discharge for an \code{aem} or \code{element} object
at the given x and y coordinates.

\code{\link[=discharge]{discharge()}} computes the \verb{x, y} and \code{z} components of the discharge vector for an \code{aem} or \code{element} object
at the given x, y and z coordinates.

\code{\link[=darcy]{darcy()}} computes the \verb{x, y} and \code{z} components of the Darcy flux vector (also called specific discharge vector)
for an \code{aem} object at the given x, y and z coordinates.

\code{\link[=velocity]{velocity()}} computes the \verb{x, y} and \code{z} components of the average linear groundwater flow velocity vector
for an \code{aem} object at the given x, y and z coordinates.
}
\details{
There is no \code{\link[=darcy]{darcy()}} or \code{\link[=velocity]{velocity()}} method for an object of class \code{element} because an \code{aem} object is required
to calculate the saturated thickness using \code{\link[=satthick]{satthick()}}.
}
\examples{

w <- well(xw = 50, yw = 0, Q = 200)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = 100)
ml <- aem(k = 10, top = 10, base = 0, n = 0.2, w, uf)

domega(ml, c(50, 0), c(25, -25))

xg <- seq(-100, 100, length = 500)
yg <- seq(-75, 75, length = 100)
domega(ml, xg, yg, as.grid = TRUE)

discharge(ml, c(50, 0), c(25, -25), magnitude = TRUE)
discharge(ml, xg, yg, as.grid = TRUE)

darcy(ml, c(50, 0), c(25, -25), magnitude = TRUE)
darcy(ml, xg, yg, as.grid = TRUE)

velocity(ml, c(50, 0), c(25, -25), magnitude = TRUE, R = 5)
velocity(ml, xg, yg, as.grid = TRUE, R = 5)

# For elements
domega(w, c(50, 0), c(-25, 25))

discharge(w, c(50, 0), c(-25, 25))
discharge(w, c(50, 0), c(-25, 25), as.grid = TRUE, magnitude = TRUE)

}
\seealso{
\code{\link[=state-variables]{state-variables()}}, \code{\link[=satthick]{satthick()}}, \code{\link[=head_to_potential]{head_to_potential()}}
}
