% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aem.R
\name{head.aem}
\alias{head.aem}
\title{Calculate the hydraulic head}
\usage{
\method{head}{aem}(aem, x, y, as.grid = FALSE, ...)
}
\arguments{
\item{aem}{\code{aem} object}

\item{x}{numeric x coordinates to evaluate \code{head} at}

\item{y}{numeric y coordinates to evaluate \code{head} at}

\item{as.grid}{logical, should a matrix of dimensions c(\verb{length(y), length(x)}) be returned? Defaults to \code{FALSE}.}

\item{...}{ignored}
}
\value{
A vector of \code{length(x)} (equal to \code{length(y)}) with the hydraulic head values at \code{x} and \code{y}.
If \code{as.grid = TRUE}, a matrix of dimensions c(\verb{length(y), length(x)}) described by
marginal vectors \code{x} and \code{y} containing the hydraulic head values at the grid points.
}
\description{
\code{\link[=head.aem]{head.aem()}} computes the hydraulic head at the given x and y coordinates.
}
\examples{
TR <- 100
w <- well(xw = 50, yw = 0, Q = 200)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = TR)
ml <- aem(TR = 100, w, uf)

head(ml, c(50, 0), c(25, -25))

xg <- seq(-100, 100, length = 500)
yg <- seq(-75, 75, length = 100)
head(ml, xg, yg, as.grid = TRUE)

}
