% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linesink.R
\name{headlinesink}
\alias{headlinesink}
\title{Create a head-specified line-sink analytic element}
\usage{
headlinesink(TR, x0, y0, x1, y1, hc, ...)
}
\arguments{
\item{TR}{numeric, transmissivity value of aquifer}

\item{x0}{numeric, starting x location of line-sink}

\item{y0}{numeric, starting y location of line-sink}

\item{x1}{numeric, ending x location of line-sink}

\item{y1}{numeric, ending y location of line-sink}

\item{hc}{numeric, specified hydraulic head of the line-sink}

\item{...}{ignored}
}
\value{
Head-specified line-sink analytic element which is an object of class \code{headlinesink} and inherits from \code{linesink}.
}
\description{
\code{\link[=headlinesink]{headlinesink()}} creates a line-sink analytic element with constant specified head. The discharge
into the line-sink per unit length is computed by solving the corresponding \code{aem} model.
}
\examples{
headlinesink(TR = 100, -75, 50, 100, 50, hc = 10)

}
\seealso{
\code{\link[=linesink]{linesink()}}
}
