% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{headwell}
\alias{headwell}
\title{Create a analytic element of a well with a constant head}
\usage{
headwell(TR, xw, yw, hw, rw = 0.3, ...)
}
\arguments{
\item{TR}{numeric, transmissivity value of aquifer}

\item{xw}{numeric, x location of the well}

\item{yw}{numeric, y location of the well}

\item{hw}{numeric, specified hydraulic head in the well}

\item{rw}{numeric, radius of the well. Defaults to 0.3 (meter).}

\item{...}{ignored}
}
\value{
Analytic element of a well with constant head which is an object of class \code{headwell} and inherits from \code{well}.
}
\description{
\code{\link[=headwell]{headwell()}} creates an analytic element of a well with a constant, specified head. The discharge
into the well is computed by solving the corresponding \code{aem} model.
}
\examples{
headwell(xw = 400, yw = 300, hw = 20, rw = 0.3, TR = 100)

}
