% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linedoublet.R
\name{linedoublet}
\alias{linedoublet}
\title{Create a line-doublet analytic element with a specified resistance}
\usage{
linedoublet(x0, y0, x1, y1, resistance, ...)
}
\arguments{
\item{x0}{numeric, starting x location of line-doublet}

\item{y0}{numeric, starting y location of line-doublet}

\item{x1}{numeric, ending x location of line-doublet}

\item{y1}{numeric, ending y location of line-doublet}

\item{resistance}{numeric, hydraulic resistance of the line-doublet}

\item{...}{ignored}
}
\value{
Resistance-specified line-doublet analytic element which is an object of class \code{linedoublet} and inherits from \code{element}.
}
\description{
\code{\link[=linedoublet]{linedoublet()}} creates a line-doublet analytic element with a constant specified resistance
}
\details{
The resistance is used to calculate the strength of the line-doublet at the collocation points.
If the aquifer is unconfined (i.e. has a variable saturated thickness), the system of equations becomes
non-linear with respect to the hydraulic head and iteration is required to solve the model.

For the special case that \code{resistance = Inf}, the line-doublet models an impermeable wall.

Three collocation points are defined per line-doublet: one at each vertex of the line-doublet and one in
the center of the line. A single line-doublet therefore adds three equations to the system of equations.
The strength distribution is defined at these collocation points and is approximated by a parabolically varying
strength distribution in between.
}
\examples{

# linedoublet(-75, 50, 100, 50, res = 500) # TODO uncomment when line-doublets are fixed

# Impermeable wall
# linedoublet(-75, 50, 100, 50, resistance = Inf) # TODO uncomment when line-doublets are fixed

}
