% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R, R/aem.R, R/element.R
\name{omega}
\alias{omega}
\alias{omega.aem}
\alias{omega.element}
\title{Calculate the complex potential}
\usage{
omega(...)

\method{omega}{aem}(aem, x, y, as.grid = FALSE, ...)

\method{omega}{element}(element, x, y, ...)
}
\arguments{
\item{...}{ignored}

\item{aem}{\code{aem} object}

\item{x}{numeric x coordinates to evaluate \code{omega} at}

\item{y}{numeric y coordinates to evaluate \code{omega} at}

\item{as.grid}{logical, should a matrix of dimensions c(\verb{length(y), length(x)}) be returned? Defaults to \code{FALSE}.}

\item{element}{analytic element of class \code{element}}
}
\value{
A vector of \code{length(x)} (equal to \code{length(y)}) with the complex potential values at \code{x} and \code{y}.
If \code{as.grid = TRUE}, a matrix of dimensions \code{c(length(y), length(x))} described by
marginal vectors \code{x} and \code{y} containing the complex potential values at the grid points.
}
\description{
\code{\link[=omega]{omega()}} computes the complex potential for an \code{aem} or \code{element} object
at the given x and y coordinates.
}
\examples{

w <- well(xw = 50, yw = 0, Q = 200)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = 100)
ml <- aem(k = 10, top = 10, base = 0, n = 0.2, w, uf)

omega(ml, c(50, 0), c(25, -25))

xg <- seq(-100, 100, length = 500)
yg <- seq(-75, 75, length = 100)
omega(ml, xg, yg, as.grid = TRUE)
w <- well(xw = 50, yw = 0, Q = 200)
omega(w, c(50, 0), c(-25, 25))

}
\seealso{
\code{\link[=heads]{heads()}}, \code{\link[=potential]{potential()}}, \code{\link[=streamfunction]{streamfunction()}}, \code{\link[=domega]{domega()}}, \code{\link[=discharge]{discharge()}}
}
