% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.aem}
\alias{plot.aem}
\title{Plot the locations of all elements in an \code{aem} object with point or line geometries}
\usage{
\method{plot}{aem}(aem, add = FALSE, xlim, ylim, frame.plot = TRUE, ...)
}
\arguments{
\item{aem}{\code{aem} object}

\item{add}{logical, should the plot be added to the existing plot? Defaults to \code{FALSE}.}

\item{xlim}{numeric, plot limits along the x-axis. Required if \code{add = FALSE}.}

\item{ylim}{numeric, plot limits along the y-axis. Required if \code{add = FALSE}.}

\item{frame.plot}{logical, should a border be drawn around the plot. Defaults to \code{TRUE}.}

\item{...}{additional arguments passed to \code{\link[=plot]{plot()}}.}
}
\value{
Plot of the planar location of all analytic elements, or added to the existing plot.
}
\description{
\code{\link[=plot.aem]{plot.aem()}} plots the planar locations of all analytic elements with a point or line geometry
in an \code{aem} object, or adds them to an existing plot.
}
\examples{
w <- well(xw = 50, yw = 0, Q = 200)
wi <- well(xw = 0, yw = 0, Q = -100)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = 100)
ls <- linesink(-75, 50, 100, 50, sigma = 1)
ml <- aem(k = 10, top = 10, base = 0, n = 0.2, w, wi, uf, ls)

plot(ml, xlim = c(-100, 100), ylim = c(-75, 75))

xg <- seq(-100, 100, length = 500)
yg <- seq(-75, 75, length = 100)

contour(ml, xg, yg, nlevels = 20, col = 'dodgerblue3')
plot(ml, add = TRUE)

}
