% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.element}
\alias{plot.element}
\title{Plot the location of an analytic element with a point or line geometry}
\usage{
\method{plot}{element}(element, add = FALSE, pch = 16, cex = 0.75, ...)
}
\arguments{
\item{element}{analytic element of class \code{element} to plot. If not a point or line geometry, nothing is plotted.}

\item{add}{logical, should the plot be added to the existing plot? Defaults to \code{FALSE}.}

\item{pch}{numeric point symbol value, defaults to \code{16}.}

\item{cex}{numeric symbol size value, defaults to \code{0.75}.}

\item{...}{additional arguments passed to \code{\link[=plot]{plot()}}}
}
\value{
Plot of the planar location of the element, if any, or added to the existing plot.
}
\description{
\code{\link[=plot.element]{plot.element()}} plots the planar location of an analytic element with a point
or line geometry or adds it to an existing plot.
}
\examples{
ls <- linesink(x0 = -200, y0 = -150, x1 = 200, y1 = 150, sigma = 0.1)
w <- well(xw = 50, yw = 0, Q = 200)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = 100)

plot(ls)
plot(w, add = TRUE)
plot(uf) # empty

}
