% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R, R/aem.R, R/element.R
\name{potential}
\alias{potential}
\alias{potential.aem}
\alias{potential.element}
\title{Calculate the discharge potential}
\usage{
potential(...)

\method{potential}{aem}(aem, x, y, as.grid = FALSE, ...)

\method{potential}{element}(element, x, y, ...)
}
\arguments{
\item{...}{ignored}

\item{aem}{\code{aem} object}

\item{x}{numeric x coordinates to evaluate \code{potential} at}

\item{y}{numeric y coordinates to evaluate \code{potential} at}

\item{as.grid}{logical, should a matrix of dimensions c(\verb{length(y), length(x)}) be returned? Defaults to \code{FALSE}.}

\item{element}{analytic element of class \code{element}}
}
\value{
A vector of \code{length(x)} (equal to \code{length(y)}) with the discharge potential values at \code{x} and \code{y}.
If \code{as.grid = TRUE}, a matrix of dimensions \code{c(length(y), length(x))} described by
marginal vectors \code{x} and \code{y} containing the discharge potential values at the grid points.
}
\description{
\code{\link[=potential]{potential()}} computes the discharge potential for an \code{aem} or \code{element} object
at the given x and y coordinates.
}
\examples{
TR <- 100
w <- well(xw = 50, yw = 0, Q = 200)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = TR)
ml <- aem(TR = 100, w, uf)

potential(ml, c(50, 0), c(25, -25))

xg <- seq(-100, 100, length = 500)
yg <- seq(-75, 75, length = 100)
potential(ml, xg, yg, as.grid = TRUE)

w <- well(xw = 50, yw = 0, Q = 200)
potential(w, c(50, 0), c(-25, 25))

}
\seealso{
\code{\link[=head.aem]{head.aem()}}, \code{\link[=omega]{omega()}}, \code{\link[=streamfunction]{streamfunction()}}, \code{\link[=domega]{domega()}}, \code{\link[=discharge]{discharge()}}
}
