% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracelines.R
\name{reached_line}
\alias{reached_line}
\title{Check if particle has reached a line element}
\usage{
reached_line(line, x, y, ...)
}
\arguments{
\item{line}{analytic element of class \code{linesink}}

\item{x}{x-coordinate of particle}

\item{y}{y-coordinate of particle}

\item{...}{arguments passed to \code{\link[=point_on_line]{point_on_line()}}}
}
\value{
logical indicating if particle is on line
}
\description{
Check if particle has reached a line element
}
\details{
This assumes a fully penetrating line segment over the entire aquifer thickness.
}
\examples{
ls <- headlinesink(-75, 50, 100, 50, 10)
reached_line(ls, 50, 20)
reached_line(ls, 50, 50)

}
