% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aem.R
\name{solve.aem}
\alias{solve.aem}
\title{Solve an \code{aem} model}
\usage{
\method{solve}{aem}(aem, ...)
}
\arguments{
\item{aem}{\code{aem} object}

\item{...}{ignored}
}
\value{
The solved \code{aem} object, i.e. after finding the solution
to the system of equations as constructed by the contained elements.
}
\description{
\code{\link[=solve.aem]{solve.aem()}} solves system of equations as constructed by the supplied  elements in the \code{aem} model
}
\details{
In order to be solved, \code{aem} should contain at least
1 \code{headequation} element.
\code{\link[=solve.aem]{solve.aem()}} sets up the system of equations, and calls \code{\link[=solve]{solve()}} to
solve.
Constructing an \code{aem} object by a call to \code{\link[=aem]{aem()}} automatically calls \code{\link[=solve.aem]{solve.aem()}}.
}
\examples{

rf <- constant(-500, 0, 20)
hdw <- headwell(xw = 0, yw = 100, rw = 0.3, hw = 8)
ml <- aem(k = 10, top = 10, base = 0, n = 0.2) |>
           add_element(rf, name = 'rf') |>
           add_element(hdw, name = 'headwell')
ml <- solve(ml)

}
