% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R, R/aem.R
\name{streamfunction}
\alias{streamfunction}
\alias{streamfunction.aem}
\title{Calculate the streamfunction}
\usage{
streamfunction(...)

\method{streamfunction}{aem}(aem, x, y, as.grid = FALSE, ...)
}
\arguments{
\item{...}{ignored}

\item{aem}{\code{aem} object}

\item{x}{numeric x coordinates to evaluate \code{streamfunction} at}

\item{y}{numeric y coordinates to evaluate \code{streamfunction} at}

\item{as.grid}{logical, should a matrix of dimensions c(\verb{length(y), length(x)}) be returned? Defaults to \code{FALSE}.}
}
\value{
A vector of \code{length(x)} (equal to \code{length(y)}) with the streamfunction values at \code{x} and \code{y}.
If \code{as.grid = TRUE}, a matrix of dimensions \code{c(length(y), length(x))} described by
marginal vectors \code{x} and \code{y} containing the streamfunction values at the grid points.
}
\description{
\code{\link[=streamfunction]{streamfunction()}} computes the streamfunction for an \code{aem} or \code{element} object
at the given x and y coordinates.
}
\examples{
TR <- 100
w <- well(xw = 50, yw = 0, Q = 200)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = TR)
ml <- aem(TR = 100, w, uf)

streamfunction(ml, c(50, 0), c(25, -25))

xg <- seq(-100, 100, length = 500)
yg <- seq(-75, 75, length = 100)
streamfunction(ml, xg, yg, as.grid = TRUE)

}
\seealso{
\code{\link[=head.aem]{head.aem()}}, \code{\link[=omega]{omega()}}, \code{\link[=potential]{potential()}}, \code{\link[=disc]{disc()}}, \code{\link[=disvec]{disvec()}}
}
