% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform-flow.R
\name{uniformflow}
\alias{uniformflow}
\title{Create a uniform-flow analytic element with specified recharge}
\usage{
uniformflow(TR, gradient, angle, ...)
}
\arguments{
\item{TR}{numeric, transmissivity value of aquifer}

\item{gradient}{numeric, hydraulic gradient. Positive in the direction of flow.}

\item{angle}{numeric, angle of the primary direction of background flow,
in degrees counterclockwise from the x-axis.}

\item{...}{ignored}
}
\value{
Analytic element of constant uniform flow which is an object of class \code{uf} and inherits from \code{element}.
}
\description{
\code{\link[=uniformflow]{uniformflow()}} creates a analytic element of constant uniform background flow.
}
\examples{
uniformflow(gradient = 0.002, angle = -45, TR = 100) # South-eastern direction
}
