% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aem.R
\name{aem}
\alias{aem}
\title{Create a analytic element model}
\usage{
aem(TR, ...)
}
\arguments{
\item{TR}{numeric, transmissivity value of aquifer}

\item{...}{objects of class \code{element}, or a single list with \code{element} objects}
}
\value{
Object of class \code{aem} which is a list consisting of \code{TR}, a list containing all
elements, and a logical \code{solved} indicating if the model is solved.
}
\description{
\code{\link[=aem]{aem()}} creates an analytic element model to which elements can be added
}
\details{
If an element of class \code{headequation} is supplied, \code{solved} is called on the \code{aem}
object before it is returned.
}
\examples{
TR <- 100
w <- well(xw = 50, yw = 0, Q = 200)
rf <- constant(TR, xc = -500, yc = 0, h = 20)
uf <- uniformflow(gradient = 0.002, angle = -45, TR = TR)

aem(TR, w, rf, uf)
aem(TR, list(w, rf, uf))

}
