% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracelines.R, R/plot.R
\name{capzone}
\alias{capzone}
\alias{plot.capzone}
\title{Calculate capture zone of a well element}
\usage{
capzone(aem, well, time, npar = 30, dt = time/100, zstart = aem$base, ...)

\method{plot}{capzone}(
  x,
  y = NULL,
  as.poly = TRUE,
  col = ifelse(as.poly, "#BEBEBE90", "black"),
  add = FALSE,
  ...
)
}
\arguments{
\item{aem}{\code{aem} object}

\item{well}{well analytic element of class \code{well} or inherits from it.}

\item{time}{numeric, time of the capture zone}

\item{npar}{integer, number of particles to use in the backward tracking. Defaults to 30.}

\item{dt}{numeric, time discretization used in the particle tracking. Defaults \code{time / 100}.}

\item{zstart}{numeric value with the starting elevation of the particles. Defaults to the base of the aquifer.}

\item{...}{additional arguments passed to \code{\link[=tracelines]{tracelines()}} for \code{\link[=capzone]{capzone()}} or additional arguments passed to \code{\link[=polygon]{polygon()}} or \code{\link[=plot.tracelines]{plot.tracelines()}} when plotting.}

\item{x}{object of class \code{capzone}}

\item{y}{ignored}

\item{as.poly}{logical, should the polygon of the capture zone be plotted as drawn from the endpoints (\code{TRUE}, default)
or should the tracelines be plotted using \code{\link[=plot.tracelines]{plot.tracelines()}} (\code{FALSE})?}

\item{col}{colour of polygon or tracelines. Defaults to a light grey with reduced opacity when \code{as.poly = TRUE} and \code{'black'} otherwise.}

\item{add}{logical, should the plot be added to the existing plot? Defaults to \code{FALSE}.}
}
\value{
\code{\link[=capzone]{capzone()}} returns an object of class \code{capzone} which is simply the output of \code{\link[=tracelines]{tracelines()}} with an additional class
used for calling the plot method.
}
\description{
\code{\link[=capzone]{capzone()}} determines the capture zone of a well element in the flow field by performing backward
particle tracking until the requested time is reached.

\code{\link[=plot.capzone]{plot.capzone()}} plots a \code{capzone} object as a polygon or draws the tracelines.
}
\details{
\code{\link[=capzone]{capzone()}} is a thin wrapper around \code{\link[=tracelines]{tracelines()}}. Backward particle tracking is performed using \code{\link[=tracelines]{tracelines()}}
and setting \code{forward = FALSE}. Initial particle locations are computed by equally spacing \code{npar} locations at the well
radius at the \code{zstart} elevation. To obtain a sharper delineation of the capture zone, try using more particles.

Note that different \code{zstart} values only have an effect in models with vertical flow components.
}
\examples{

k <- 10
top <- 10; base <- 0
n <- 0.3

uf <- uniformflow(TR = 100, gradient = 0.001, angle = -10)
rf <- constant(TR, xc = -1000, yc = 0, hc = 20)
w1 = well(200, 50, Q = 250)
w2 = well(-200, -100, Q = 450)

m <- aem(k, top, base, n = n, uf, rf, w1, w2)

cp5 <- capzone(m, w1, time = 5*365)
cp10 <- capzone(m, w2, time = 10*365, as.poly = FALSE)

xg <- seq(-800, 800, length = 100)
yg <- seq(-500, 500, length = 100)

contours(m, xg, yg, col = 'dodgerblue3', nlevels = 20)
plot(cp5, add = TRUE)
plot(cp10, add = TRUE, as.poly = FALSE) # plot tracelines instead of polygon

# model with vertical flow components
as <- areasink(0, 0, N = 0.001, R = 1500)
m <- aem(k, top, base, n = n, uf, rf, w1, w2, as)

# two different starting levels
cp5a <- capzone(m, w1, time = 5*365, zstart = base)
cp5b <- capzone(m, w1, time = 5*365, zstart = 8)

contours(m, xg, yg, col = 'dodgerblue3', nlevels = 20)
plot(cp5a, add = TRUE)
plot(cp5b, add = TRUE, col = 'orange3') # smaller zone

}
\seealso{
\code{\link[=tracelines]{tracelines()}}
}
