% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linedoublet.R
\name{linedoublet}
\alias{linedoublet}
\title{Create a strength-specified line-doublet analytic element}
\usage{
linedoublet(x0, y0, x1, y1, resistance, ...)
}
\arguments{
\item{x0}{numeric, starting x location of line-doublet}

\item{y0}{numeric, starting y location of line-doublet}

\item{x1}{numeric, ending x location of line-doublet}

\item{y1}{numeric, ending y location of line-doublet}

\item{resistance}{numeric, hydraulic resistance of the line-doublet}

\item{...}{ignored}
}
\value{
Strength-specified line-doublet analytic element which is an object of class \code{linedoublet} and inherits from \code{element}.
}
\description{
\code{\link[=linedoublet]{linedoublet()}} creates a line-doublet analytic element with constant specified strength.
}
\details{
The resistance is used to calculate the strength of the line-doublet. If the aquifer is
unconfined (i.e. has a variable saturated thickness), the system of equations becomes non-linear
with respect to the hydraulic head and iteration is required to solve the model.
}
\examples{
linedoublet(-75, 50, 100, 50, res = 500)

}
